requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.6.16
chardet==3.0.4
Django==2.1.11
django-celery==3.3.0
django-celery-beat==1.5.0
django-timezone-field==3.0
idna==2.8
kombu==3.0.37
PyMySQL==0.9.3
python-crontab==2.3.7
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
PyAutoGUI==0.9.53
